HUUUUGE Refactoring
